requirements.txt
PyYAML
anthropic>=0.39.0
archon-ai
cerebras-cloud-sdk>=1.0.0
datasets==3.6.0
einops
google-generativeai>=0.8.3
groq
hjson
huggingface-hub
loguru
matplotlib
modal
multiprocess
ninja
numpy==1.26
openai==1.75.0
packaging
pandas
parse_llm_code
protobuf==5.29.4
pydra_config
pytest
rank-bm25
regex
requests
tenacity
together
torch==2.5.0
tqdm
transformers
triton==3.1.0
